# dennic_admin_api_gateway/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Dennic Admin Project",
        "contact": {},
        "version": "1.7"
    },
    "host": "admin.dennic.uz",
    "paths": {
...
        "/v1/doctor-notes": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "ListDoctorNotes - API to list doctor notes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Doctor Note"
                ],
                "summary": "ListDoctorNotes",
                "parameters": [
                    {
                        "enum": [
                            "prescription"
                        ],
                        "type": "string",
                        "description": "searchField",
                        "name": "searchField",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "is_active",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model_booking_service.DoctorNotesType"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "UpdateDoctorNote - API to update a doctor note",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Doctor Note"
                ],
                "summary": "UpdateDoctorNote",
                "parameters": [
                    {
                        "description": "UpdateDoctorNoteReq",
                        "name": "UpdateDoctorNoteReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model_booking_service.UpdateDoctorNoteReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model_booking_service.DoctorNote"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "DeleteDoctorNote - API to delete a doctor note",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Doctor Note"
                ],
                "summary": "DeleteDoctorNote",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "hard_delete",
                        "name": "hard_delete",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StatusRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    }
                }
            }
        },
        "/v1/doctor-notes/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "CreateDoctorNote - Api for create doctor note",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Doctor Note"
                ],
                "summary": "CreateDoctorNote",
                "parameters": [
                    {
                        "description": "CreateDoctorNotesReq",
                        "name": "CreateDoctorNotesReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model_booking_service.CreateDoctorNotesReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model_booking_service.DoctorNote"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    }
                }
            }
        },
        "/v1/doctor-notes/get": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "GetDoctorNote - API to get doctor note by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Doctor Note"
                ],
                "summary": "GetDoctorNote",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "is_active",
                        "name": "is_active",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model_booking_service.DoctorNote"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model_booking_service.CreateDoctorNotesReq": {
            "type": "object",
            "properties": {
                "appointment_id": {
                    "type": "integer"
                },
                "doctor_id": {
                    "type": "string"
                },
                "patient_id": {
                    "type": "string"
                },
                "prescription": {
                    "type": "string"
                }
            }
        },
...
        "model_booking_service.DoctorNote": {
            "type": "object",
            "properties": {
                "appointment_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "doctor_id": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "patient_id": {
                    "type": "string"
                },
                "prescription": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model_booking_service.DoctorNotesType": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "doctor_notes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model_booking_service.DoctorNote"
                    }
                }
            }
        },
...
        "model_booking_service.UpdateDoctorNoteReq": {
            "type": "object",
            "properties": {
                "appointment_id": {
                    "type": "integer"
                },
                "doctor_id": {
                    "type": "string"
                },
                "doctor_notes_id": {
                    "type": "string"
                },
                "patient_id": {
                    "type": "string"
                },
                "prescription": {
                    "type": "string"
                }
            }
        },
...
        "model_common.ResponseError": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "model_common.StandardErrorModel": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/model_common.ResponseError"
                }
            }
        },
...
        "models.StatusRes": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}